Extract shared concurrent prefill/decode dispatch helper
Mooncake, P2P (OffloadingConnector), and SGLang each reimplemented the
same concurrent-dispatch pattern nearly line for line: an async prefill
goroutine with panic recovery, a synchronous decode on the calling
goroutine, and the same OTel span/timing bookkeeping around both.
Extract it into a single runConcurrentPD helper and have each connector
supply only its protocol-specific request bodies and connector name.

Also fixes SGLang's inline status-code check to use the existing
isHTTPError helper instead of a hand-rolled range comparison.
