+++ pkg/sidecar/proxy/connector_mooncake.go
@@ -25,12 +25,7 @@ import (
 	"net/http"
 	"time"
 
-	"go.opentelemetry.io/otel/attribute"
-	"go.opentelemetry.io/otel/codes"
-	"go.opentelemetry.io/otel/trace"
-
 	"github.com/llm-d/llm-d-router/pkg/common/observability/logging"
-	"github.com/llm-d/llm-d-router/pkg/common/observability/tracing"
 	reqcommon "github.com/llm-d/llm-d-router/pkg/common/request"
 )
 
@@ -132,7 +127,11 @@ func (s *Server) handleMooncake(w http.ResponseWriter, r *http.Request, prefillP
 		trace.Info("Decode request", "body", string(decodeBody))
 	}
 
-	s.handleMooncakeConcurrentRequests(w, r, prefillBody, decodeBody, prefillPodHostPort, dpRank)
+	s.runConcurrentPD(w, r, prefillBody, decodeBody, prefillPodHostPort, KVConnectorMooncake, func(prefillReq, _ *http.Request) {
+		// Route prefill to the same DP rank whose engine_id was given to decode, so the
+		// KV it produces lands on the engine decode pulls from. No-op for a single rank.
+		prefillReq.Header.Set(mooncakeDataParallelRankHeader, dpRank)
+	})
 }
 
 // getMooncakeEngineMap returns the dp_rank -> engine_id mapping for the given prefill, querying the bootstrap server on first use and caching it.
@@ -182,102 +181,4 @@ func (s *Server) getMooncakeEngineMap(ctx context.Context, prefillHostPort, boot
 	// add the full map into LRU
 	s.mooncakeEngineIDs.Add(prefillHostPort, engineMap)
 	return engineMap, nil
-}
-
-func (s *Server) handleMooncakeConcurrentRequests(w http.ResponseWriter, r *http.Request, prefillBody, decodeBody []byte, prefillHost, dpRank string) {
-	tracer := tracing.Tracer(tracerScope)
-	ctx := r.Context()
-
-	// WithoutCancel for prefill so it isn't aborted when the decode response finishes first
-	prefillReq := cloneRequestWithBody(context.WithoutCancel(ctx), r, prefillBody)
-	decodeReq := cloneRequestWithBody(ctx, r, decodeBody)
-
-	// Route prefill to the same DP rank whose engine_id was given to decode, so the
-	// KV it produces lands on the engine decode pulls from. No-op for a single rank.
-	prefillReq.Header.Set(mooncakeDataParallelRankHeader, dpRank)
-
-	// Prefill runs in a goroutine: only populates KV cache, response is discarded.
-	// Decode runs on the main thread: writes the actual response back to the client via w.
-	ctx, prefillSpan := tracer.Start(ctx, "prefill",
-		trace.WithSpanKind(trace.SpanKindInternal),
-	)
-	prefillSpan.SetAttributes(
-		attribute.String("llm_d.pd_proxy.prefill_target", prefillHost),
-		attribute.String("llm_d.pd_proxy.connector", KVConnectorMooncake),
-		attribute.Bool("llm_d.pd_proxy.prefill.async", true),
-	)
-	prefillStart := time.Now()
-
-	prefillHandler, err := s.prefillerProxyHandler(prefillHost)
-	if err != nil {
-		prefillSpan.SetStatus(codes.Error, "failed to create prefill handler")
-		prefillSpan.End()
-		if err := errorBadGateway(err, w); err != nil {
-			s.logger.Error(err, "failed to send error response to client")
-		}
-		return
-	}
-
-	go func() {
-		defer prefillSpan.End()
-		defer func() {
-			if rec := recover(); rec != nil && rec != http.ErrAbortHandler {
-				s.logger.Error(fmt.Errorf("panic: %v", rec), "panic in prefill request")
-			}
-		}()
-		// buffered writer captures response for status check only, not sent to client
-		pw := &bufferedResponseWriter{}
-		prefillHandler.ServeHTTP(pw, prefillReq)
-		prefillDuration := time.Since(prefillStart)
-		prefillSpan.SetAttributes(
-			attribute.Int("llm_d.pd_proxy.prefill.status_code", pw.statusCode),
-			attribute.Float64("llm_d.pd_proxy.prefill.duration_ms", float64(prefillDuration.Milliseconds())),
-		)
-		if isHTTPError(pw.statusCode) {
-			prefillSpan.SetStatus(codes.Error, "prefill request failed")
-		}
-		s.logger.V(logging.TRACE).Info("mooncake prefill request completed", "status", pw.statusCode)
-	}()
-
-	// Decode Stage
-	ctx, decodeSpan := tracer.Start(ctx, "decode",
-		trace.WithSpanKind(trace.SpanKindInternal),
-	)
-	defer decodeSpan.End()
-
-	decodeSpan.SetAttributes(
-		attribute.String("llm_d.pd_proxy.connector", KVConnectorMooncake),
-		attribute.Bool("llm_d.pd_proxy.decode.concurrent_with_prefill", true),
-	)
-	decodeStart := time.Now()
-
-	decodeReq = decodeReq.WithContext(ctx)
-	s.decoderProxy.ServeHTTP(w, decodeReq)
-
-	decodeDuration := time.Since(decodeStart)
-	decodeSpan.SetAttributes(
-		attribute.Float64("llm_d.pd_proxy.decode.duration_ms", float64(decodeDuration.Milliseconds())),
-		attribute.String("llm_d.pd_proxy.decode.target", s.config.DecoderURL.Host),
-	)
-
-	// Calculate end-to-end P/D timing metrics for concurrent P/D.
-	// True TTFT captures time from gateway request start to decode start.
-	// In Mooncake's concurrent mode, prefill duration is tracked in the async prefill span.
-	if currentSpan := trace.SpanFromContext(ctx); currentSpan.SpanContext().IsValid() {
-		var totalDuration time.Duration
-		var trueTTFT time.Duration
-		if requestStartValue := ctx.Value(requestStartTimeKey); requestStartValue != nil {
-			if requestStart, ok := requestStartValue.(time.Time); ok {
-				totalDuration = time.Since(requestStart)
-				trueTTFT = decodeStart.Sub(requestStart)
-			}
-		}
-
-		currentSpan.SetAttributes(
-			attribute.Float64("llm_d.pd_proxy.total_duration_ms", float64(totalDuration.Milliseconds())),
-			attribute.Float64("llm_d.pd_proxy.true_ttft_ms", float64(trueTTFT.Milliseconds())),
-			attribute.Float64("llm_d.pd_proxy.decode_duration_ms", float64(decodeDuration.Milliseconds())),
-			attribute.Bool("llm_d.pd_proxy.concurrent_pd", true),
-		)
-	}
 }
