+++ pkg/sidecar/proxy/connector_p2p.go
@@ -17,7 +17,6 @@ limitations under the License.
 package proxy
 
 import (
-	"context"
 	"encoding/json"
 	"fmt"
 	"io"
@@ -27,14 +26,8 @@ import (
 	"os"
 	"strconv"
 	"strings"
-	"time"
 
-	"go.opentelemetry.io/otel/attribute"
-	"go.opentelemetry.io/otel/codes"
-	"go.opentelemetry.io/otel/trace"
-
-	"github.com/llm-d/llm-d-router/pkg/common/observability/logging"
-	"github.com/llm-d/llm-d-router/pkg/common/observability/tracing"
+	logging "github.com/llm-d/llm-d-router/pkg/common/observability/logging"
 	reqcommon "github.com/llm-d/llm-d-router/pkg/common/request"
 )
 
@@ -118,99 +111,7 @@ func (s *Server) handleP2P(w http.ResponseWriter, r *http.Request, prefillPodHos
 		v.Info("decode request body", "body", string(decodeBody))
 	}
 
-	s.handleP2PConcurrentRequests(w, r, prefillBody, decodeBody, prefillPodHostPort)
-}
-
-func (s *Server) handleP2PConcurrentRequests(w http.ResponseWriter, r *http.Request, prefillBody, decodeBody []byte, prefillHost string) {
-	tracer := tracing.Tracer(tracerScope)
-	ctx := r.Context()
-
-	// WithoutCancel for prefill so it isn't aborted when the decode response finishes first.
-	prefillReq := cloneRequestWithBody(context.WithoutCancel(ctx), r, prefillBody)
-	decodeReq := cloneRequestWithBody(ctx, r, decodeBody)
-
-	// Prefill runs in a goroutine: only stores KV, response is discarded.
-	// Decode runs on the main thread: writes the actual response back via w.
-	ctx, prefillSpan := tracer.Start(ctx, "prefill",
-		trace.WithSpanKind(trace.SpanKindInternal),
-	)
-	prefillSpan.SetAttributes(
-		attribute.String("llm_d.pd_proxy.prefill_target", prefillHost),
-		attribute.String("llm_d.pd_proxy.connector", KVConnectorOffloading),
-		attribute.Bool("llm_d.pd_proxy.prefill.async", true),
-	)
-	prefillStart := time.Now()
-
-	prefillHandler, err := s.prefillerProxyHandler(prefillHost)
-	if err != nil {
-		prefillSpan.SetStatus(codes.Error, "failed to create prefill handler")
-		prefillSpan.End()
-		if err := errorBadGateway(err, w); err != nil {
-			s.logger.Error(err, "failed to send error response to client")
-		}
-		return
-	}
-
-	go func() {
-		defer prefillSpan.End()
-		defer func() {
-			if rec := recover(); rec != nil && rec != http.ErrAbortHandler {
-				s.logger.Error(fmt.Errorf("panic: %v", rec), "panic in prefill request")
-			}
-		}()
-		pw := &bufferedResponseWriter{}
-		prefillHandler.ServeHTTP(pw, prefillReq)
-		prefillDuration := time.Since(prefillStart)
-		prefillSpan.SetAttributes(
-			attribute.Int("llm_d.pd_proxy.prefill.status_code", pw.statusCode),
-			attribute.Float64("llm_d.pd_proxy.prefill.duration_ms", float64(prefillDuration.Milliseconds())),
-		)
-		if isHTTPError(pw.statusCode) {
-			prefillSpan.SetStatus(codes.Error, "prefill request failed")
-		}
-		s.logger.V(logging.DEBUG).Info("p2p prefill request completed", "status", pw.statusCode)
-	}()
-
-	// Decode Stage
-	ctx, decodeSpan := tracer.Start(ctx, "decode",
-		trace.WithSpanKind(trace.SpanKindInternal),
-	)
-	defer decodeSpan.End()
-
-	decodeSpan.SetAttributes(
-		attribute.String("llm_d.pd_proxy.connector", KVConnectorOffloading),
-		attribute.Bool("llm_d.pd_proxy.decode.concurrent_with_prefill", true),
-	)
-	decodeStart := time.Now()
-
-	decodeReq = decodeReq.WithContext(ctx)
-	s.decoderProxy.ServeHTTP(w, decodeReq)
-
-	decodeDuration := time.Since(decodeStart)
-	decodeSpan.SetAttributes(
-		attribute.Float64("llm_d.pd_proxy.decode.duration_ms", float64(decodeDuration.Milliseconds())),
-		attribute.String("llm_d.pd_proxy.decode.target", s.config.DecoderURL.Host),
-	)
-
-	// End-to-end P/D timing. True TTFT captures time from gateway request start
-	// to decode start; prefill duration is tracked in the async prefill span.
-	if currentSpan := trace.SpanFromContext(ctx); currentSpan.SpanContext().IsValid() {
-		var totalDuration time.Duration
-		var trueTTFT time.Duration
-		if requestStartValue := ctx.Value(requestStartTimeKey); requestStartValue != nil {
-			if requestStart, ok := requestStartValue.(time.Time); ok {
-				totalDuration = time.Since(requestStart)
-				trueTTFT = decodeStart.Sub(requestStart)
-			}
-		}
-
-		currentSpan.SetAttributes(
-			attribute.Float64("llm_d.pd_proxy.total_duration_ms", float64(totalDuration.Milliseconds())),
-			attribute.Float64("llm_d.pd_proxy.true_ttft_ms", float64(trueTTFT.Milliseconds())),
-			attribute.Float64("llm_d.pd_proxy.decode_duration_ms", float64(decodeDuration.Milliseconds())),
-			attribute.Bool("llm_d.pd_proxy.concurrent_pd", true),
-		)
-	}
+	s.runConcurrentPD(w, r, prefillBody, decodeBody, prefillPodHostPort, KVConnectorOffloading, nil)
 }
 
 // p2pPullAvailable reports whether this deployment can pull cached prefix over
